Fix leftjoins in listings.service.ts Refactor seeds module
